Align numpy-operator dtype conventions with PyTorch
Where MXNet's numpy frontend previously followed NumPy and PyTorch
diverges, follow PyTorch (plus project overrides):

- int x float binary promotion keeps the float operand's width
  (int32 x float16 -> float16, not NumPy's float64). This also fixes the
  mixed-precision backward crash: the output now always equals the float
  input, so the cast-both compute path is never taken.
- result_type: float-width for arrays; Python scalars are weak
  (int32 array + python float -> float32).
- linalg.norm / var / std / average over integer input -> float32.
- unique(float16) accepted (CPU + GPU).
- integer cross accepted (forward type switch widened to
  MSHADOW_TYPE_SWITCH; sign-flip ternary made type-stable for half/bf16).
- Migrate legacy take/ravel/unravel tests to integer indices; update
  promotion/result_type/norm/unique/cross tests to assert PyTorch
  conventions.

Document remaining integer-dtype gaps in issues.md: INT1 (mean(int) ->
int32 rounded) and INT2 (integer matmul/dot/tensordot). Both need real
kernel work (fused integer-mean reduce / integer GEMM); naive cast
fallbacks hit kTempSpace collision + GPU async use-after-free, so they
are left rejected.

+++ src/operator/numpy/linalg/np_norm.cc
@@ -122,7 +122,9 @@ bool NumpyNormType(const nnvm::NodeAttrs& attrs,
   int in_type = in_attrs->at(0);
   int out_type;
   if (!common::is_float(in_type)) {
-    out_type = mshadow::kFloat64;
+    // PyTorch errors on integer norm; project convention returns float32
+    // (NumPy returns float64).
+    out_type = mshadow::kFloat32;
   } else {
     out_type = in_type;
   }
